// packages/lib/src/enums.d.ts
export declare enum PageType {
    FOLDER = "FOLDER",
    DOCUMENT = "DOCUMENT",
    CHANNEL = "CHANNEL",
    AI_CHAT = "AI_CHAT",
    CANVAS = "CANVAS",
    FILE = "FILE",
    SHEET = "SHEET"
}
export declare enum PermissionAction {
    VIEW = "VIEW",
    EDIT = "EDIT",
    SHARE = "SHARE",
    DELETE = "DELETE"
}
export declare enum SubjectType {
    USER = "USER"
}
//# sourceMappingURL=enums.d.ts.map